+++ app/src/main/java/com/example/myapplication/ui/home/viewmodel/HomeViewModel.java
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.home;
+package com.example.myapplication.ui.home.viewmodel;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
